add query_contracts.go and tests
